+++ discord-bot/bot/src/main/java/fr/pacifista/bot/discord/PacifistaDiscordBotApp.java
