+++ discord-bot/bot/src/main/resources/application.properties
