+++ discord-bot/modules/core-discord/src/main/java/fr/pacifista/bot/discord/modules/core/utils/Colors.java
@@ -1,0 +1,12 @@
+package fr.pacifista.bot.discord.modules.core.utils;
+
+import lombok.NoArgsConstructor;
+
+import java.awt.*;
+
+@NoArgsConstructor(access = lombok.AccessLevel.PRIVATE)
+public class Colors {
+
+    public static final Color PACIFISTA_COLOR = Color.decode("#2cafff");
+
+}
