+++ discord-bot/modules/giveaways-discord/pom.xml
@@ -4,20 +4,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>fr.pacifista.bot</groupId>
-        <artifactId>pacifista-bot</artifactId>
+        <groupId>fr.pacifista.bot.discord.modules</groupId>
+        <artifactId>pacifista-bot-discord-modules</artifactId>
         <version>2.0.0</version>
-        <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <groupId>fr.pacifista.bot.discord</groupId>
-    <artifactId>pacifista-bot-discord</artifactId>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>modules</module>
-        <module>bot</module>
-    </modules>
+    <groupId>fr.pacifista.bot.discord.modules.giveaway</groupId>
+    <artifactId>giveaways-discord</artifactId>
 
     <properties>
         <maven.compiler.source>${java.version}</maven.compiler.source>
@@ -27,14 +20,8 @@
 
     <dependencies>
         <dependency>
-            <groupId>net.dv8tion</groupId>
-            <artifactId>JDA</artifactId>
-            <version>${discord.jda.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>fr.pacifista.bot.core</groupId>
-            <artifactId>core</artifactId>
+            <groupId>fr.pacifista.bot.discord.modules.core</groupId>
+            <artifactId>pacifista-bot-discord-core</artifactId>
             <version>${fr.pacifista.bot.version}</version>
         </dependency>
     </dependencies>
