+++ discord-bot/modules/pom.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>fr.pacifista.bot.discord</groupId>
+        <artifactId>pacifista-bot-discord</artifactId>
+        <version>2.0.0</version>
+    </parent>
+
+    <groupId>fr.pacifista.bot.discord.modules</groupId>
+    <artifactId>pacifista-bot-discord-modules</artifactId>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>core-discord</module>
+        <module>support-tickets-discord</module>
+        <module>giveaways-discord</module>
+    </modules>
+
+    <properties>
+        <maven.compiler.source>${java.version}</maven.compiler.source>
+        <maven.compiler.target>${java.version}</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
+
+</project>
