--- discord-bot/src/main/java/fr/pacifista/bot/discord/events/buttons/Button.java
@@ -1,8 +1,0 @@
-package fr.pacifista.bot.discord.events.buttons;
-
-import lombok.NonNull;
-import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
-
-public interface Button {
-    void onButton(@NonNull ButtonInteractionEvent event);
-}
