--- discord-bot/src/main/java/fr/pacifista/bot/discord/utils/Colors.java
@@ -1,12 +1,0 @@
-package fr.pacifista.bot.discord.utils;
-
-import lombok.NoArgsConstructor;
-
-import java.awt.*;
-
-@NoArgsConstructor(access = lombok.AccessLevel.PRIVATE)
-public class Colors {
-
-    public static final Color PACIFISTA_COLOR = Color.decode("#2cafff");
-
-}
